Moved UserInfoFragment to be under ui
